requirements.txt
flask==2.0.1
python-dotenv==0.19.0
openai==1.3.0
PyPDF2==3.0.1
langchain==0.0.350
chromadb==0.4.18
tiktoken==0.5.2
langchain_community
langchain-openai
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
unstructured